rename messages package to message
