+++ Dockerfile
@@ -12,7 +12,7 @@ COPY apis/ apis/
 COPY config/ config/
 COPY routes/ routes/
 COPY presenter/ presenter/
-COPY messages/ messages/
+COPY message/ message/
 COPY repositories/ repositories/
 
 RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -a -o cfapi main.go
